reorganize callback module structure
