+++ docs/source/_rst/callback/processing/metric_tracker.rst
@@ -1,0 +1,7 @@
+Metric Tracker
+==================
+.. currentmodule:: pina.callback.processing.metric_tracker
+
+.. autoclass:: MetricTracker
+   :members:
+   :show-inheritance:
