+++ docs/source/_rst/callback/processing/pina_progress_bar.rst
@@ -1,0 +1,7 @@
+PINA Progress Bar
+==================
+.. currentmodule:: pina.callback.processing.pina_progress_bar
+
+.. autoclass:: PINAProgressBar
+   :members:
+   :show-inheritance:
