--- docs/source/_rst/callback/processing_callback.rst
@@ -1,11 +1,0 @@
-Processing callbacks
-=======================
-
-.. currentmodule:: pina.callback.processing_callback
-.. autoclass:: MetricTracker
-   :members:
-   :show-inheritance:
-
-.. autoclass:: PINAProgressBar
-   :members:
-   :show-inheritance:
