+++ pina/callback/__init__.py
@@ -3,14 +3,15 @@
 __all__ = [
     "SwitchOptimizer",
     "SwitchScheduler",
-    "MetricTracker",
-    "PINAProgressBar",
-    "R3Refinement",
     "NormalizerDataCallback",
+    "PINAProgressBar",
+    "MetricTracker",
+    "R3Refinement",
 ]
 
-from .optimizer_callback import SwitchOptimizer
-from .processing_callback import MetricTracker, PINAProgressBar
+from .optim.switch_optimizer import SwitchOptimizer
+from .optim.switch_scheduler import SwitchScheduler
+from .processing.normalizer_data_callback import NormalizerDataCallback
+from .processing.pina_progress_bar import PINAProgressBar
+from .processing.metric_tracker import MetricTracker
 from .refinement import R3Refinement
-from .normalizer_data_callback import NormalizerDataCallback
-from .switch_scheduler import SwitchScheduler
