+++ tests/test_callback/test_switch_optimizer.py
